+++ lib/vagrant/errors.rb
@@ -184,10 +184,6 @@ module Vagrant
       error_key(:box_not_found_with_provider)
     end
 
-    class BoxOutdatedNoBox < VagrantError
-      error_key(:box_outdated_no_box)
-    end
-
     class BoxProviderDoesntMatch < VagrantError
       error_key(:box_provider_doesnt_match)
     end