+++ CHANGELOG.md
@@ -22,6 +22,8 @@ BUG FIXES:
   - core: Prerelease plugins installed are locked to that prerelease
       version. [GH-3301]
   - core: Better error message when adding a box with a malformed version. [GH-3332]
+  - core: Fix a rare issue where vagrant up would complain it couldn't
+      check version of a box that doesn't exist. [GH-3326]
   - commands/box: Show versions when listing. [GH-3316]
   - commands/status: Machine readable output contains the target. [GH-3218]
   - guests/arch: Reload udev rules after network change. [GH-3322]